pyproject.toml
[tool.poetry]
name = "fullend"
version = "0.1.0"
description = ""
authors = ["zglian <f74102030@gs.ncku.edu.tw>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.10.4"
fastapi = "^0.110.0"
psycopg2-binary = "^2.9.6"
sqlalchemy = "^2.0.28"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
psycopg2 = "^2.9.6"
pygame = "^2.5.2"
pytz = "^2024.1"
pathlib = "^1.0.1"
opencv-python = "^4.9.0.80"
azure-storage-blob = "^12.19.1"
azure-identity = "^1.15.0"
aiohttp = "^3.9.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
